functions/requirements.txt
azure-core==1.30.2
azure-cosmos==4.7.0
azure-functions==1.20.0
azure-servicebus
azure-storage-blob==12.20.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
common @ git+https://github.com/it-minds/itm_leads-finder-scripts.git@main#subdirectory=internal
cryptography==42.0.8
idna==3.7
isodate==0.6.1
langchain-core==0.2.3
pycparser==2.22
pydantic==2.7.2
requests==2.32.3
six==1.16.0
soupsieve==2.5
tenacity==8.3.0
typing_extensions==4.12.1
urllib3==2.2.1
